% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SubSaharanAfrica_Pf_incidence}
\alias{SubSaharanAfrica_Pf_incidence}
\title{Plasmodium falciparum parasite rate in 2-10 year olds in 2000-2017 in Sub-Saharan Africa}
\format{A data frame with 144000 rows and 5 variables:
\describe{
  \item{x}{Longitude in decimal degrees}
  \item{y}{Latitude in decimal degrees}
  \item{Median}{The median posterior predictive rate of Plasmodium falciparum in 2-10 year olds, 2000-2017}
  \item{UCI}{The lower extreme of the credible interval around the median posterior predictive rate of Plasmodium falciparum in 2-10 year olds, 2000-2017}
  \item{UCI}{The upper extreme of the credible interval around the median posterior predictive rate of Plasmodium falciparum in 2-10 year olds, 2000-2017}
}}
\source{
\url{https://map.ox.ac.uk/}
}
\usage{
SubSaharanAfrica_Pf_incidence
}
\description{
A dataset containing longitude, latitude, point estimates, and credible intervals bounds
}
\keyword{datasets}
