% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SubSaharanAfrica_shp}
\alias{SubSaharanAfrica_shp}
\title{Shape files for Sub-Saharan Africa}
\format{A SpatialPolygonsDataFrame with dimensions 59 and 16
\describe{
  \item{iso}{To-be-completed}
  \item{admn_level}{To-be-completed}
}}
\source{
Obtained using malariaAtlas::getShp (see data-raw/get_shape_files.R)
}
\usage{
SubSaharanAfrica_shp
}
\description{
A SpatialPolygonsDataFrame for Sub-Saharan Africa
}
\keyword{datasets}
