% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CentralAfrica_Pf_incidence}
\alias{CentralAfrica_Pf_incidence}
\title{\emph{Plasmodium falciparum} predicted incidence}
\format{
An data frame with 270083 observations and four variables:
\describe{
  \item{x}{Longitude in decimal degrees}
  \item{y}{Latitude in decimal degrees}
  \item{z}{Median predicted incidence at location x y}
  \item{u}{Width of the 95\% predicted incidence credible interval at location x y}
}
}
\source{
These data are available at the Malaria Atlas Project (MAP) website \url{https://map.ox.ac.uk/}.
   Specifically, they were obtained by selecting
   'ANNUAL MEAN OF PF INCIDENCE' at \url{https://map.ox.ac.uk/malaria-burden-data-download/}.
}
\usage{
CentralAfrica_Pf_incidence
}
\description{
\emph{P. falciparum} predicted all-age incidence (clinical cases per 1,000 population per annum)
in 2017 for central Africa at 2.5 arcminute (approximately 5km) resolution [1].
}
\details{
The median and credible interval were computed using samples from a posterior
predictive simulation that approximated the joint posterior predictive
distribution thereby accounting for spatial covariance [1,2].
}
\examples{
str(CentralAfrica_Pf_incidence)
head(CentralAfrica_Pf_incidence)
}
\references{
\describe{
  \item{[1]}{Weiss DJ, Lucas TCD, Nguyen M, et al. Mapping the global
  prevalence, incidence, and mortality of Plasmodium falciparum, 2000--17: a
  spatial and temporal modelling study. Lancet 2019.}
  \item{[2]}{Gething PW, Patil AP, and Hay SI. Quantifying aggregated
  uncertainty in Plasmodium falciparum malaria prevalence and populations at
  risk via efficient space-time geostatistical joint simulation. PLoS
  Computational Biology 2010.}
}
}
\keyword{datasets}
