% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SubSaharanAfrica_Pf_incidence}
\alias{SubSaharanAfrica_Pf_incidence}
\title{\emph{Plasmodium falciparum} predicted incidence rate}
\format{An data frame with a row per prediction (obs) and four variables:
\describe{
  \item{x}{Longitude in decimal degrees}
  \item{y}{Latitude in decimal degrees}
  \item{z}{The median of a ‘per-pixel’ posterior predictive simulation of \emph{P. falciparum} incidence rate in 2017}
  \item{u}{The width of the 95\% credible interval of a ‘per-pixel’ posterior predictive simulation of \emph{P. falciparum}
  incidence rate in 2017.}
}}
\source{
These data are available at the Malaria Atlas Project (MAP) website \url{https://map.ox.ac.uk/}.
   Specifically, they were obtained by selecting
   'ANNUAL MEAN OF PF INCIDENCE' at \url{https://map.ox.ac.uk/malaria-burden-data-download/}.
}
\usage{
SubSaharanAfrica_Pf_incidence
}
\description{
\emph{P. falciparum} predicted all-age incidence rate (clinical cases per 1,000 population per annum)
in 2017 for sub-Saharan Africa at 2.5 arcminute (approximately 5km) resolution.
}
\examples{
str(SubSaharanAfrica_Pf_incidence)
head(SubSaharanAfrica_Pf_incidence)
}
\references{
Weiss DJ, Lucas TCD, Nguyen M, et al.
   Mapping the global prevalence, incidence, and mortality of Plasmodium falciparum, 2000--17:
   a spatial and temporal modelling study. Lancet 2019; published online June 19.
}
\keyword{datasets}
